requirements.txt
click==8.1.3
colorama==0.4.6
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
plotly==5.13.0
PyYAML==6.0
tenacity==8.2.1
Werkzeug==2.2.2
